requirements.txt
ansicolors
pyperclip